#338: registration confirmation has a link to the confirmation page, plus I did a little more commenting and refactoring
