+++ uber/custom_tags.py
@@ -431,13 +431,13 @@ class organization_or_event_name(template.Node):
 class single_day_prices(template.Node):
     def render(self, context):
         prices = ''
-        for day, price in conf['badge_prices']['single_day'].items():
+        for day, price in BADGE_PRICES['single_day'].items():
             prices += '${} for {}, '.format(price, day)
-        return prices + 'and ${} for other days'.format(conf['badge_prices']['default_single_day'])
+        return prices + 'and ${} for other days'.format(BADGE_PRICES['default_single_day'])
 
 class Notice(template.Node):
     def notice(self, label, takedown, discount=False):
-        discount = conf['badge_prices']['group_discount'] if discount else 0
+        discount = BADGE_PRICES['group_discount'] if discount else 0
         for day, price in sorted(PRICE_BUMPS.items()):
             if localized_now() < day:
                 return 'Price goes up to ${} at 11:59pm EST on {}'.format(price - discount, (day - timedelta(days=1)).strftime('%A, %b %e'))
