+++ uber/decorators.py
@@ -138,7 +138,7 @@ def renderable_data(data=None):
     data = data or {}
     data['PAGE'] = cherrypy.request.path_info.split('/')[-1]
     data.update({m.__name__: m for m in Session.all_models()})  # TODO: I don't think we actually need this anymore
-    data.update({k: v for k,v in constants.__dict__.items() if re.match('^[_A-Z0-9]*$', k)})
+    data.update({k: v for k,v in config.__dict__.items() if re.match('^[_A-Z0-9]*$', k)})
     data.update({k: getattr(state, k) for k in dir(state) if re.match('^[_A-Z0-9]*$', k)})
     for date in DATES:
         before, after = 'BEFORE_' + date, 'AFTER_' + date
@@ -153,7 +153,7 @@ def renderable_data(data=None):
 
     access = AdminAccount.access_set()
     for acctype in ['ACCOUNTS','PEOPLE','STUFF','MONEY','CHALLENGES','CHECKINS']:
-        data['HAS_' + acctype + '_ACCESS'] = getattr(constants, acctype) in access
+        data['HAS_' + acctype + '_ACCESS'] = getattr(config, acctype) in access
 
     return data
