+++ uber/templates/emails/reg_workflow/attendee_confirmation.html
@@ -8,10 +8,19 @@ You ({{ attendee.full_name }}) have preregistered for {{ EVENT_NAME }}
 {% else %}
     and your payment of ${{ attendee.amount_paid }} has been received.  
 {% endif %}
-Your registration confirmation number is {{ attendee.id }}.
+Your registration confirmation number is {{ attendee.id }}, and you can update your information or kick in
+{% if attendee.amount_extra %}
+    even more
+{% endif %}
+extra money <a href="{{ URL_BASE }}/preregistration/confirm?id={{ attendee.id }}">here</a>.
 
 <br/> <br/>
-Badges are not mailed out before the event, so your badge will be available for pickup at the registration desk when you arrive at {{ EVENT_NAME }}. Simply bring a photo ID to the registration desk, where you'll be provided with your badge. If you ordered bonus items such as a ribbon, t-shirt, or supporter package, you can pick those up at our merchandise booth. The location and hours of the registration desk and merchandise booth will be e-mailed prior to the event.
+
+Badges are not mailed out before the event, so your badge will be available for pickup at the registration desk when you arrive at {{ EVENT_NAME }}.  Simply bring a photo ID to the registration desk, where you'll be provided with your badge.
+{% if DONATIONS_ENABLED %}
+    If you ordered bonus items such as a ribbon, t-shirt, or supporter package, you can pick those up at our merchandise booth.
+    The location and hours of the registration desk and merchandise booth will be e-mailed prior to the event.
+{% endif %}
 
 {% if attendee.amount_extra %}
     <br/> <br/>
